refactor: create DescriptionWithChildren component and remove deprecated typography components
